Finish "Notifications" and "Home" pages
I added a personal message from me in the "Notifications" section. I also created some fake spam messages to fill out the page. I added a post linking to my calculator so that all of my Github projects are contained within this website. There are a few elements that are not fully finished but that I plan to revisit later, so I turned off their display for now. Lastly, I made some small textual changes to posts in the "Trending" section.